ta-bot_backend/Remove_Idle_Conversations/requirements.txt
pymongo==4.5.0
python-dotenv==1.0.0
redis==5.0.1

